Grounded-Segment-Anything/requirements.txt
addict==2.4.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
filelock==3.13.1
Flask==3.1.0
fonttools==4.57.0
fsspec==2024.6.1
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@126abe633ffe333e16e4a0a4e946bc1003caf757#egg=groundingdino&subdirectory=GroundingDINO
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.0.0
platformdirs==4.3.7
pycocotools==2.0.8
pyparsing==3.2.3
PyRep==4.1.0.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@126abe633ffe333e16e4a0a4e946bc1003caf757#egg=segment_anything&subdirectory=segment_anything
six==1.17.0
supervision==0.21.0
sympy==1.13.3
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.67.1
transformers==4.51.3
triton==2.2.0
typing_extensions==4.12.2
urllib3==2.4.0
Werkzeug==3.1.3
yapf==0.43.0
